# be/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yummy API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "List reservations based on user role",
                "operationId": "fbc6050d9f7fdb07075d3aaeb157ffa9",
                "responses": {
                    "200": {
                        "description": "Reservations retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Create a reservation for the current user",
                "operationId": "01e3d21dc330a31de025203bc142ffa3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "starts_at",
                                    "guests_count"
                                ],
                                "properties": {
                                    "starts_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-08-01T18:00:00"
                                    },
                                    "guests_count": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Window seat"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created"
                    },
                    "403": {
                        "description": "Only guests can create reservations"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reservations/user": {
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Admin creates reservation for a registered user",
                "operationId": "ddbcbdcf73e43979bff1e8ad3fba7570",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "starts_at",
                                    "guests_count"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "starts_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-08-01T18:00:00"
                                    },
                                    "guests_count": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Birthday dinner"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created for user"
                    },
                    "403": {
                        "description": "Only admins can access this"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reservations/phone": {
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Admin creates reservation for an unregistered guest by phone",
                "operationId": "f1995d49ecca719488de2ea25bbd0060",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "last_name",
                                    "starts_at",
                                    "guests_count"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+421912345678"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "Anna"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Smith"
                                    },
                                    "starts_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-08-01T19:00:00"
                                    },
                                    "guests_count": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Allergic to peanuts"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created for unregistered guest"
                    },
                    "403": {
                        "description": "Only admins can access this"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reservations/{id}": {
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Cancel a reservation (soft delete)",
                "operationId": "9b6641e4ded1ff7adfce352d589b85e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reservation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation cancelled"
                    },
                    "403": {
                        "description": "Only admins can cancel reservations"
                    },
                    "404": {
                        "description": "Reservation not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a token returned from the login endpoint",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}